requirements.txt
aiohttp==0.21.6
cchardet==1.0.0
python-Levenshtein==0.12.0